+++ packages/app/app/components/Resource/Webpage/Editor.vue
@@ -1,31 +1,13 @@
 <script setup lang="ts">
-import type { Editor } from "grapesjs";
-
 import { GRAPES_JS_EDITOR_CONTAINER_ID, SURVEY_INVITE_BLOCK_CATEGORY } from "@/services/grapesjs/constants";
 import { setBlocks } from "@/services/grapesjs/setBlocks";
 import { createWebpageSurveyInviteBlocks } from "@/services/webpageEditor/createWebpageSurveyInviteBlocks";
+import { WebpageEditorPlugins } from "@/services/webpageEditor/WebpageEditorPlugins";
+import { WebpageEditorStyleManager } from "@/services/webpageEditor/WebpageEditorStyleManager";
 import { useWebpageEditorStore } from "@/store/webpageEditor";
 import { ResourceType } from "@esposter/db-schema";
-import { usePlugin } from "grapesjs";
-import grapesJSBlocksBasic from "grapesjs-blocks-basic";
-import grapesJSComponentCountdown from "grapesjs-component-countdown";
-import grapesJSCustomCode from "grapesjs-custom-code";
-import grapesJSParserPostcss from "grapesjs-parser-postcss";
 import "grapesjs/dist/css/grapes.min.css";
-import grapesJSPluginExport from "grapesjs-plugin-export";
-import grapesJSPluginForms from "grapesjs-plugin-forms";
-import grapesJSPresetWebpage from "grapesjs-preset-webpage";
-import grapesJSStyleBg from "grapesjs-style-bg";
-// @ts-expect-error no d.ts file
-import grapesJSTabs from "grapesjs-tabs";
-import grapesJSTooltip from "grapesjs-tooltip";
-// @ts-expect-error no d.ts file
-import grapesJSTouch from "grapesjs-touch";
-import grapesJSTuiImageEditor from "grapesjs-tui-image-editor";
-import grapesJSTyped from "grapesjs-typed";
-import jsBeautify from "js-beautify";
 
-const { css: cssFormat, html: htmlFormat } = jsBeautify;
 const webpageEditorStore = useWebpageEditorStore();
 const { readWebpageEditor, saveWebpageEditor } = webpageEditorStore;
 const uploadFile = useUploadResourceFile(ResourceType.Webpage, () => webpageEditorStore.resource?.id ?? "");
@@ -36,336 +18,10 @@ const { editor } = await useGrapesJsEditor(
     store: (data, storeEditor) => saveWebpageEditor(data, { css: storeEditor.getCss(), html: storeEditor.getHtml() }),
   },
   {
-    plugins: [
-      usePlugin(grapesJSBlocksBasic, {
-        flexGrid: true,
-      }),
-      grapesJSPluginForms,
-      grapesJSComponentCountdown,
-      grapesJSPluginExport,
-      usePlugin(grapesJSTabs, {
-        tabsBlock: { category: "Extra" },
-      }),
-      grapesJSCustomCode,
-      grapesJSTouch,
-      grapesJSParserPostcss,
-      grapesJSTooltip,
-      grapesJSTuiImageEditor,
-      usePlugin(grapesJSTyped, {
-        block: {
-          category: "Extra",
-          content: {
-            strings: ["Text row one", "Text row two", "Text row three"],
-            type: "typed",
-            "type-speed": 40,
-          },
-        },
-      }),
-      grapesJSStyleBg,
-      usePlugin(grapesJSPresetWebpage, {
-        modalImportContent: (editorInstance: Editor) => {
-          const html = editorInstance.getHtml();
-          const css = editorInstance.getCss();
-          return css ? `${htmlFormat(html)}<style>\n${cssFormat(css)}</style>` : htmlFormat(html);
-        },
-        modalImportLabel: "<div text-title-small mb-2.5>Paste here your HTML/CSS and click Import</div>",
-        modalImportTitle: "Import Template",
-      }),
-    ],
+    plugins: WebpageEditorPlugins,
     selectorManager: { componentFirst: true },
     showOffsets: true,
-    styleManager: {
-      sectors: [
-        {
-          name: "General",
-          properties: [
-            {
-              default: "none",
-              extend: "float",
-              options: [
-                { className: "fa fa-times", id: "times", value: "none" },
-                { className: "fa fa-align-left", id: "left", value: "left" },
-                { className: "fa fa-align-right", id: "right", value: "right" },
-              ],
-              type: "radio",
-            },
-            "display",
-            { extend: "position", type: "select" },
-            "top",
-            "right",
-            "left",
-            "bottom",
-          ],
-        },
-        {
-          name: "Dimension",
-          open: false,
-          properties: [
-            "width",
-            {
-              id: "flex-width",
-              name: "Width",
-              property: "flex-basis",
-              toRequire: true,
-              type: "integer",
-              units: ["px", "%"],
-            },
-            "height",
-            "max-width",
-            "min-height",
-            "margin",
-            "padding",
-          ],
-        },
-        {
-          name: "Typography",
-          open: false,
-          properties: [
-            "font-family",
-            "font-size",
-            "font-weight",
-            "letter-spacing",
-            "color",
-            "line-height",
-            {
-              extend: "text-align",
-              options: [
-                { className: "fa fa-align-left", id: "left", label: "Left" },
-                { className: "fa fa-align-center", id: "center", label: "Center" },
-                { className: "fa fa-align-right", id: "right", label: "Right" },
-                { className: "fa fa-align-justify", id: "justify", label: "Justify" },
-              ],
-            },
-            {
-              default: "none",
-              options: [
-                { className: "fa fa-times", id: "none", label: "None" },
-                { className: "fa fa-underline", id: "underline", label: "underline" },
-                { className: "fa fa-strikethrough", id: "line-through", label: "Line-through" },
-              ],
-              property: "text-decoration",
-              type: "radio",
-            },
-            "text-shadow",
-          ],
-        },
-        {
-          name: "Decorations",
-          open: false,
-          properties: ["opacity", "border-radius", "border", "box-shadow", "background"],
-        },
-        {
-          buildProps: ["transition", "perspective", "transform"],
-          name: "Extra",
-          open: false,
-        },
-        {
-          name: "Flex",
-          open: false,
-          properties: [
-            {
-              defaults: "block",
-              list: [
-                { id: "disable", name: "Disable", value: "block" },
-                { id: "enable", name: "Enable", value: "flex" },
-              ],
-              name: "Flex Container",
-              property: "display",
-              type: "select",
-            },
-            {
-              name: "Flex Parent",
-              property: "label-parent-flex",
-              type: "integer",
-            },
-            {
-              defaults: "row",
-              list: [
-                {
-                  className: "icons-flex icon-dir-row",
-                  id: "row",
-                  name: "Row",
-                  title: "Row",
-                  value: "row",
-                },
-                {
-                  className: "icons-flex icon-dir-row-rev",
-                  id: "row-reverse",
-                  name: "Row reverse",
-                  title: "Row reverse",
-                  value: "row-reverse",
-                },
-                {
-                  className: "icons-flex icon-dir-col",
-                  id: "column",
-                  name: "Column",
-                  title: "Column",
-                  value: "column",
-                },
-                {
-                  className: "icons-flex icon-dir-col-rev",
-                  id: "column-reverse",
-                  name: "Column reverse",
-                  title: "Column reverse",
-                  value: "column-reverse",
-                },
-              ],
-              name: "Direction",
-              property: "flex-direction",
-              type: "radio",
-            },
-            {
-              defaults: "flex-start",
-              list: [
-                {
-                  className: "icons-flex icon-just-start",
-                  id: "flex-start",
-                  title: "Start",
-                  value: "flex-start",
-                },
-                {
-                  className: "icons-flex icon-just-end",
-                  id: "flex-end",
-                  title: "End",
-                  value: "flex-end",
-                },
-                {
-                  className: "icons-flex icon-just-sp-bet",
-                  id: "space-between",
-                  title: "Space between",
-                  value: "space-between",
-                },
-                {
-                  className: "icons-flex icon-just-sp-ar",
-                  id: "space-around",
-                  title: "Space around",
-                  value: "space-around",
-                },
-                {
-                  className: "icons-flex icon-just-sp-cent",
-                  id: "center",
-                  title: "Center",
-                  value: "center",
-                },
-              ],
-              name: "Justify",
-              property: "justify-content",
-              type: "radio",
-            },
-            {
-              defaults: "center",
-              list: [
-                {
-                  className: "icons-flex icon-al-start",
-                  id: "flex-start",
-                  title: "Start",
-                  value: "flex-start",
-                },
-                {
-                  className: "icons-flex icon-al-end",
-                  id: "flex-end",
-                  title: "End",
-                  value: "flex-end",
-                },
-                {
-                  className: "icons-flex icon-al-str",
-                  id: "stretch",
-                  title: "Stretch",
-                  value: "stretch",
-                },
-                {
-                  className: "icons-flex icon-al-center",
-                  id: "center",
-                  title: "Center",
-                  value: "center",
-                },
-              ],
-              name: "Align",
-              property: "align-items",
-              type: "radio",
-            },
-            {
-              name: "Flex Children",
-              property: "label-parent-flex",
-              type: "integer",
-            },
-            {
-              defaults: "0",
-              min: 0,
-              name: "Order",
-              property: "order",
-              type: "integer",
-            },
-            {
-              name: "Flex",
-              properties: [
-                {
-                  defaults: "0",
-                  min: 0,
-                  name: "Grow",
-                  property: "flex-grow",
-                  type: "integer",
-                },
-                {
-                  defaults: "0",
-                  min: 0,
-                  name: "Shrink",
-                  property: "flex-shrink",
-                  type: "integer",
-                },
-                {
-                  defaults: "auto",
-                  name: "Basis",
-                  property: "flex-basis",
-                  type: "integer",
-                  unit: "",
-                  units: ["px", "%", ""],
-                },
-              ],
-              property: "flex",
-              type: "composite",
-            },
-            {
-              defaults: "auto",
-              list: [
-                {
-                  id: "auto",
-                  name: "Auto",
-                  value: "auto",
-                },
-                {
-                  className: "icons-flex icon-al-start",
-                  id: "flex-start",
-                  title: "Start",
-                  value: "flex-start",
-                },
-                {
-                  className: "icons-flex icon-al-end",
-                  id: "flex-end",
-                  title: "End",
-                  value: "flex-end",
-                },
-                {
-                  className: "icons-flex icon-al-str",
-                  id: "stretch",
-                  title: "Stretch",
-                  value: "stretch",
-                },
-                {
-                  className: "icons-flex icon-al-center",
-                  id: "center",
-                  title: "Center",
-                  value: "center",
-                },
-              ],
-              name: "Align",
-              property: "align-self",
-              type: "radio",
-            },
-          ],
-        },
-      ],
-    },
+    styleManager: WebpageEditorStyleManager,
   },
   { upload: uploadFile },
 );
